@@ -2,37 +2,42 @@
   "name": "fullapp",
   "version": "1.0.0",
   "license": "UNLICENSED",
+  "type": "module",
   "engines": {
     "node": ">=16"
   },
   "scripts": {
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "format": "prettier --write 'app/frontend/**/*.{html,css,scss,json,js,ts,svelte}'",
-    "types": "tsc --noEmit; svelte-check"
+    "lint": "prettier --check app/frontend/ && eslint app/frontend/"
   },
   "devDependencies": {
-    "@apollo/client": "^3.3.6",
-    "@popperjs/core": "^2.6.0",
-    "@rails/actioncable": "^7.0.0",
-    "@roxi/routify": "^2.12.5",
-    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.33",
-    "@types/bootstrap": "^5.0.16",
-    "@types/qrcode": "^1.4.2",
-    "@types/rails__actioncable": "^6.1.4",
-    "bootstrap": "^5.0.0-beta1",
-    "graphql": "^16.0.0",
-    "graphql-ruby-client": "^1.8.1",
-    "postcss": "^8.4.19",
-    "postcss-preset-env": "^7.1.0",
-    "prettier": "^2.0.5",
-    "prettier-plugin-svelte": "^2.3.1",
-    "qrcode": "^1.4.4",
-    "sass": "^1.29.0",
-    "svelte": "^3.29.7",
-    "svelte-check": "^2.2.0",
-    "svelte-i18n": "^3.3.0",
-    "svelte-preprocess": "^4.7.3",
-    "typescript": "^4.3.2",
-    "vite": "^4.2.1",
-    "vite-plugin-ruby": "^3.2.1"
+    "@apollo/client": "^3.7.12",
+    "@popperjs/core": "^2.11.7",
+    "@rails/actioncable": "^7.0.4",
+    "@sveltejs/adapter-static": "^2.0.2",
+    "@sveltejs/kit": "^1.15.5",
+    "@types/bootstrap": "^5.2.6",
+    "@types/qrcode": "^1.5.0",
+    "@types/rails__actioncable": "^6.1.6",
+    "@typescript-eslint/eslint-plugin": "^5.58.0",
+    "@typescript-eslint/parser": "^5.58.0",
+    "bootstrap": "^5.2.3",
+    "eslint": "^8.38.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-svelte3": "^4.0.0",
+    "graphql": "^16.6.0",
+    "graphql-ruby-client": "^1.11.7",
+    "prettier": "^2.8.7",
+    "prettier-plugin-svelte": "^2.10.0",
+    "qrcode": "^1.5.1",
+    "sass": "^1.62.0",
+    "svelte": "^3.58.0",
+    "svelte-check": "^3.2.0",
+    "svelte-i18n": "^3.6.0",
+    "svelte-preprocess": "^5.0.3",
+    "tslib": "^2.5.0",
+    "typescript": "^5.0.4",
+    "vite": "^4.2.1"
   }
 }
