@@ -1,14 +1,9 @@
 import { defineConfig } from 'vite'
-import RubyPlugin from 'vite-plugin-ruby'
-import { svelte } from '@sveltejs/vite-plugin-svelte'
+import { sveltekit } from '@sveltejs/kit/vite'
 
-// https://vitejs.dev/config/
 export default defineConfig({
   clearScreen: false,
-  plugins: [
-    RubyPlugin(),
-    svelte({})
-  ],
+  plugins: [sveltekit()],
   css: {
     devSourcemap: true
   }
